# openapi.yaml
openapi: 3.0.3
info:
  title: High-Assurance API
  version: 1.0.0
  description: Multi-route API with Security, Outbox Logic, and Full Testing Suite.

servers:
  - url: http://localhost:5000
    description: Local Development API

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
        '503':
          description: Service degraded

  /login:
    post:
      summary: User Login (Get Token)
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT Issued
        '400':
          description: Bad request
        '401':
          description: Invalid credentials
        '429':
          description: Rate limited

  /upload-dataset:
    post:
      summary: Async Dataset Upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted for processing
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '422':
          description: Validation failed

  /transfer:
    post:
      summary: Secure Fund Transfer (Outbox Pattern)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to_user, amount]
              properties:
                to_user:
                  type: string
                  example: "senior_engineer"
                amount:
                  type: number
                  example: 5000
      responses:
        '201':
          description: Transaction Recorded in Outbox
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Account not found
        '409':
          description: Duplicate transaction
        '422':
          description: Validation failed